Retire the Dirty Unicorns Interface navigation suite (DUI) 2015-2019
It has become apparent over the last couple cycles that the concept
of the software navigation bar is being deprecated. Thus, we must
adapt accordingly. Fling and SmartBar will FOREVER live in our memories
as the greatest navitation bar solution in the history of AOSP. Pulse
lives on, for now...